+++ graphql-java-runtime/src/main/java/com/graphql_java_generator/client/SubscriptionClient.java
@@ -6,11 +6,14 @@ package com.graphql_java_generator.client;
 import com.graphql_java_generator.exception.GraphQLRequestExecutionException;
 
 /**
+ * Now useless, with the Spring reactive implementation. All interactions with the Web Socket are managed by the
+ * {@link SubscriptionCallback}.<BR/>
  * This interface allows the application to interact with the subscription, once it has subscribed to it. The currently
  * only available action is
  * 
  * @author etienne-sf
  */
+@Deprecated
 public interface SubscriptionClient {
 
 	/**