+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-Forum-client/src/main/java/com/graphql_java_generator/samples/forum/Main.java
@@ -2,6 +2,8 @@ package com.graphql_java_generator.samples.forum;
 
 import java.util.Calendar;
 import java.util.Collections;
+import java.util.Date;
+import java.util.GregorianCalendar;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
@@ -18,7 +20,14 @@ import com.graphql_java_generator.exception.GraphQLRequestPreparationException;
 import com.graphql_java_generator.samples.forum.client.Queries;
 import com.graphql_java_generator.samples.forum.client.graphql.PartialDirectRequests;
 import com.graphql_java_generator.samples.forum.client.graphql.PartialPreparedRequests;
+import com.graphql_java_generator.samples.forum.client.graphql.forum.client.GraphQLRequest;
+import com.graphql_java_generator.samples.forum.client.graphql.forum.client.Member;
+import com.graphql_java_generator.samples.forum.client.graphql.forum.client.MutationTypeExecutor;
+import com.graphql_java_generator.samples.forum.client.graphql.forum.client.Post;
+import com.graphql_java_generator.samples.forum.client.graphql.forum.client.PostInput;
 import com.graphql_java_generator.samples.forum.client.graphql.forum.client.QueryTypeExecutor;
+import com.graphql_java_generator.samples.forum.client.graphql.forum.client.SubscriptionTypeExecutor;
+import com.graphql_java_generator.samples.forum.client.graphql.forum.client.TopicPostInput;
 
 /**
  * A Spring Boot client app. Very easy to use and to configure
@@ -33,6 +42,10 @@ public class Main implements CommandLineRunner {
 	PartialDirectRequests partialDirectRequests;
 	@Autowired
 	PartialPreparedRequests partialPreparedRequests;
+	@Autowired
+	MutationTypeExecutor mutationTypeExecutor;
+	@Autowired
+	SubscriptionTypeExecutor subscriptionTypeExecutor;
 
 	public static void main(String[] args) {
 		SpringApplication.run(Main.class, args);
@@ -44,17 +57,23 @@ public class Main implements CommandLineRunner {
 	 */
 	@Override
 	public void run(String... args) throws Exception {
-		System.out.println("");
-		System.out.println("============================================================================");
-		System.out.println("======= SIMPLEST WAY: DIRECT QUERIES =======================================");
-		System.out.println("============================================================================");
-		exec(partialDirectRequests, null);
+		// System.out.println("");
+		// System.out.println("============================================================================");
+		// System.out.println("======= SIMPLEST WAY: DIRECT QUERIES =======================================");
+		// System.out.println("============================================================================");
+		// exec(partialDirectRequests, null);
+		//
+		// System.out.println("");
+		// System.out.println("============================================================================");
+		// System.out.println("======= MOST SECURE WAY: PREPARED QUERIES ==================================");
+		// System.out.println("============================================================================");
+		// exec(partialPreparedRequests, null);
 
 		System.out.println("");
 		System.out.println("============================================================================");
-		System.out.println("======= MOST SECURE WAY: PREPARED QUERIES ==================================");
+		System.out.println("======= LET'S EXECUTE A SUBSCRIPTION      ==================================");
 		System.out.println("============================================================================");
-		exec(partialPreparedRequests, null);
+		execSubscription();
 
 		System.out.println("");
 		System.out.println("");
@@ -89,6 +108,65 @@ public class Main implements CommandLineRunner {
 			System.out.println(
 					"Please start the server from the project graphql-maven-plugin-samples-StarWars-server, before executing the client part");
 		}
+	}
+
+	private void execSubscription() throws GraphQLRequestPreparationException, GraphQLRequestExecutionException {
+		// Preparation
+		GraphQLRequest subscriptionRequest = subscriptionTypeExecutor
+				.getSubscribeToNewPostGraphQLRequest("{id date author publiclyAvailable title content}");
+		GraphQLRequest createPostRequest = mutationTypeExecutor
+				.getCreatePostGraphQLRequest("{id date author{id} title content publiclyAvailable}");
+
+		PostSubscriptionCallback<Post> postSubscriptionCallback = new PostSubscriptionCallback<>(
+				Thread.currentThread());
+		Member author = new Member();
+		author.setId("12");
+		PostInput postInput = new PostInput();
+		postInput.setTopicId("22");
+		postInput.setInput(getTopicPostInput(author, "Some other content",
+				new GregorianCalendar(2020, 11 - 1, 21).getTime(), false, "The good title for a post"));
+
+		// Go, go, go
+		System.out.println("Submitting the 'subscribeToNewPostWithBindValues' GraphQL subscription");
+		subscriptionTypeExecutor.subscribeToNewPost(subscriptionRequest, postSubscriptionCallback, "Board name 1");
+
+		// Let's check that everything is ready
+		if (!postSubscriptionCallback.connected) {
+			throw new RuntimeException("The subscription should be active");
+		}
+
+		System.out.println(
+				"Creating a post (for which we expect a notification) from this postInput: " + postInput.toString());
+		mutationTypeExecutor.createPost(createPostRequest, postInput);
+
+		//////////////////////////////////////////////////////////////////////////////////////
+		// Let's check that we've received the expected notification, for this post creation
+
+		// Let's wait 10 seconds max for the post creation notification (from the subscription). We'll get interrupted
+		// before, as soon as we receive the notification
+		// (see the callback implementation in the PostSubscriptionCallback class)
+		boolean gotInterrupted = false;
+		try {
+			Thread.sleep(10 * 1000);// 10s
+		} catch (InterruptedException e) {
+			// Got interrupted (the post notification has been received)
+			gotInterrupted = true;
+		}
+		if (!gotInterrupted) {
+			throw new RuntimeException("The notification for the post creation was not received");
+		}
+	}
+
+	private TopicPostInput getTopicPostInput(Member author, String content, Date date, boolean publiclyAvailable,
+			String title) {
+		TopicPostInput input = new TopicPostInput();
+		input.setAuthorId(author.getId());
+		input.setContent(content);
+		input.setDate(date);
+		input.setPubliclyAvailable(publiclyAvailable);
+		input.setTitle(title);
+
+		return input;
 	}
 
 	@Bean