+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-Forum-server/src/main/java/org/forum/server/specific_code/DataFetchersDelegateSubscriptionTypeImpl.java
@@ -33,7 +33,7 @@ public class DataFetchersDelegateSubscriptionTypeImpl implements DataFetchersDel
 
 	@Override
 	public Publisher<Post> subscribeToNewPost(DataFetchingEnvironment dataFetchingEnvironment, String boardName) {
-		logger.debug("Received a Subscription for {}", boardName);
+		logger.info("Received a Subscription for {}", boardName);
 		return postPublisher.getPublisher(boardName);
 	}
 }